Re-Arrange some objects/modules to better fit classifications Pull Combat Detection out of core and into own sub module Pull Loading out of core and into own submodule Moved some stray icon functions out of core and into Icons.lua
