+++ DBM-Core/modules/objects/SpecRole.lua
